Refactored and splitted user agent tests into settings and on the fly.
